Refatoramento da Navbar
