+++ frontend/public/index.html
@@ -9,7 +9,7 @@
       rel="stylesheet"
     />
 
-    <title>React App</title>
+    <title>syslaudo</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
