+++ frontend/src/index.tsx
@@ -1,19 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import Home from './pages/Home';
-import PageNotFound from './pages/PageNotFound';
-import { GlobalStyle } from './style/base';
+// import { BrowserRouter, Route, Switch } from 'react-router-dom';
+// import Navbar from './components/Navbar';
+// import Home from './pages/Home';
+// import PageNotFound from './pages/PageNotFound';
+// import { GlobalStyle } from './style/base';
+import App from './App';
+
 
 ReactDOM.render(
-  <React.Fragment>
-    <GlobalStyle />
-    <BrowserRouter>
-      <Switch>
-        <Route path="/" exact={true} component={Home} />
-        <Route path="*" component={PageNotFound} />
-      </Switch>
-    </BrowserRouter>
-  </React.Fragment>,
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
   document.getElementById('root')
 );
