+++ frontend/src/pages/Home.tsx
@@ -1,8 +1,8 @@
-import Navbar from "../components/Navbar";
+// import Navbar from '../components/Navbar'
 
 function Home() {
   return (
-    <Navbar />
+    <h1>HOME</h1>
   );
 }
